Getting rid of travelling ship, updating refueling and clearing ships
Can't have a TravellingShip class because there isn't a type
column on the ship in Schemaverse.  Adding a type attr_accessor
to ship, will update this to Travelling when a ship is travelling.

Refueling ships that are Travelling only.

When a planet is surrounded only by my planets, kill all the ships
at that planet.