Adding RoundSeq, and armada groups
Checking to see if the round has changed so that I can
delete any overflow of ships that have been created.

Adding armada groups of 30 ships that all travel to the
same place to take back a planet that was lost or to a planet
that was expanded to but occupied.